# docs/openapi/pangea-backend-api.yaml
openapi: 3.1.0
info:
  title: Pangea Backend API
  version: 1.0.0
  description: |
    The Pangea Backend API provides authentication, user management, and wallet operations
    integrated with the OASIS platform.
    
    **Authentication:**
    - The API uses Better-Auth tokens for authentication (validated via JWKS)
    - All protected endpoints require a valid Better-Auth token in the `Authorization` header
    - The backend validates tokens using the frontend's JWKS endpoint (`/api/auth/jwks`)
    
    **Better-Auth Integration Flow:**
    1. User registers/logs in via Better-Auth (frontend)
    2. Frontend calls `POST /api/auth/create-oasis-avatar` with Better-Auth token
    3. Backend creates OASIS avatar and links it to the user
    4. User can now use wallet and other OASIS features
    
    **Getting Started:**
    - For Better-Auth users: Register with Better-Auth, then call `POST /api/auth/create-oasis-avatar`
    - For legacy users: Use `POST /api/auth/register` (will be deprecated)
    - After authentication, use protected endpoints with Bearer token
  contact:
    name: Pangea Backend Support
    email: support@pangea.markets
  license:
    name: Proprietary
  termsOfService: https://pangea.markets/terms
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.pangea.markets
    description: Production server
...
paths:
...
  /api/wallet/balance:
    get:
      tags:
        - Wallet
      summary: Get all wallet balances
      description: |
        Get balances for all wallets associated with the authenticated user's OASIS avatar.
        
        If no OASIS avatar exists, one will be created automatically (lazy creation).
        If no wallets exist, returns an empty array with a helpful message.
        
        Requires a valid Better-Auth token.
      operationId: getWalletBalances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balances retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalancesResponse'
              examples:
                withWallets:
                  summary: User has wallets
                  value:
                    success: true
                    balances:
                      - walletId: "wallet-123"
                        walletAddress: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
                        providerType: "SolanaOASIS"
                        balance: 100.5
                        tokenSymbol: "SOL"
                      - walletId: "wallet-456"
                        walletAddress: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                        providerType: "EthereumOASIS"
                        balance: 0.5
                        tokenSymbol: "ETH"
                    hasWallets: true
                noWallets:
                  summary: User has no wallets
                  value:
                    success: true
                    balances: []
                    message: "No wallets found. Generate a wallet first using POST /api/wallet/generate"
                    hasWallets: false
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/balance/{assetId}:
    get:
      tags:
        - Wallet
      summary: Get balance for specific asset
      description: |
        Get the balance for a specific asset/wallet.
        
        The assetId can be used to determine the blockchain (Solana or Ethereum).
        If no OASIS avatar exists, one will be created automatically (lazy creation).
        
        Requires a valid Better-Auth token.
      operationId: getAssetBalance
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          description: Asset identifier (can include blockchain hint like "solana" or "SOL")
          schema:
            type: string
            example: "solana-mainnet-SOL"
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetBalanceResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Balance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/generate:
    post:
      tags:
        - Wallet
      summary: Generate new wallet
      description: |
        Generate a new cryptocurrency wallet for the authenticated user's OASIS avatar.
        
        If no OASIS avatar exists, one will be created automatically (lazy creation).
        The wallet will be linked to the user's OASIS avatar.
        
        Requires a valid Better-Auth token.
      operationId: generateWallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateWalletRequest'
            examples:
              solana:
                summary: Generate Solana wallet
                value:
                  providerType: "SolanaOASIS"
                  setAsDefault: true
              ethereum:
                summary: Generate Ethereum wallet
                value:
                  providerType: "EthereumOASIS"
                  setAsDefault: false
      responses:
        '201':
          description: Wallet generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateWalletResponse'
        '400':
          description: Invalid provider type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/connect:
    post:
      tags:
        - Wallet
      summary: Connect external wallet
      description: |
        Connect an external wallet (Phantom, MetaMask, etc.) to the user's account.
        
        Requires wallet signature verification.
        
        Requires a valid Better-Auth token.
      operationId: connectWallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectWalletRequest'
      responses:
        '200':
          description: Wallet connected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectWalletResponse'
        '400':
          description: Invalid request or signature verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/verify:
    post:
      tags:
        - Wallet
      summary: Verify wallet ownership
      description: |
        Verify ownership of a wallet by validating a signature.
        
        Requires a valid Better-Auth token.
      operationId: verifyWallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyWalletRequest'
      responses:
        '200':
          description: Wallet verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyWalletResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/sync:
    post:
      tags:
        - Wallet
      summary: Sync wallet balances
      description: |
        Manually trigger a sync of all wallet balances for the authenticated user.
        
        If no OASIS avatar exists, one will be created automatically (lazy creation).
        
        Requires a valid Better-Auth token.
      operationId: syncWalletBalances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balances synced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBalancesResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wallet/transactions/{walletId}:
    get:
      tags:
        - Wallet
      summary: Get transaction history
      description: |
        Get transaction history for a specific wallet.
        
        The wallet must belong to the authenticated user's OASIS avatar.
        
        Requires a valid Better-Auth token.
      operationId: getWalletTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: walletId
          in: path
          required: true
          description: Wallet ID
          schema:
            type: string
            example: "wallet-123"
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Better-Auth JWT token (for Better-Auth users) or backend JWT token (for legacy users).
        
        For Better-Auth tokens:
        - Token is validated via JWKS endpoint at {FRONTEND_URL}/api/auth/jwks
        - Token must include: id, email (required), name, role, kycStatus (optional)
        
        For legacy backend tokens:
        - Token is validated using backend's JWT secret
        - Token includes: sub (user ID), email, username, avatarId, role
  schemas:
...
    WalletBalancesResponse:
      type: object
      required:
        - success
        - balances
        - hasWallets
      properties:
        success:
          type: boolean
          example: true
        balances:
          type: array
          items:
            type: object
            properties:
              walletId:
                type: string
                example: "wallet-123"
              walletAddress:
                type: string
                example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
              providerType:
                type: string
                enum: [SolanaOASIS, EthereumOASIS]
                example: "SolanaOASIS"
              balance:
                type: number
                example: 100.5
              tokenSymbol:
                type: string
                example: "SOL"
              error:
                type: string
                description: Error message if balance fetch failed
        hasWallets:
          type: boolean
          example: true
        message:
          type: string
          description: Helpful message when no wallets exist
          example: "No wallets found. Generate a wallet first using POST /api/wallet/generate"
    AssetBalanceResponse:
      type: object
      required:
        - success
        - balance
      properties:
        success:
          type: boolean
          example: true
        balance:
          type: object
          properties:
            walletId:
              type: string
              example: "wallet-123"
            walletAddress:
              type: string
              example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
            assetId:
              type: string
              example: "solana-mainnet-SOL"
            balance:
              type: number
              example: 100.5
            tokenSymbol:
              type: string
              example: "SOL"
            providerType:
              type: string
              enum: [SolanaOASIS, EthereumOASIS]
              example: "SolanaOASIS"
    GenerateWalletRequest:
      type: object
      required:
        - providerType
      properties:
        providerType:
          type: string
          enum: [SolanaOASIS, EthereumOASIS]
          description: Blockchain provider type
          example: "SolanaOASIS"
        setAsDefault:
          type: boolean
          default: true
          description: Whether to set this wallet as the default wallet
          example: true
    GenerateWalletResponse:
      type: object
      required:
        - success
        - message
        - wallet
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Wallet generated successfully for SolanaOASIS"
        wallet:
          type: object
          properties:
            walletId:
              type: string
              example: "wallet-123"
            walletAddress:
              type: string
              example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
            providerType:
              type: string
              enum: [SolanaOASIS, EthereumOASIS]
              example: "SolanaOASIS"
            isDefaultWallet:
              type: boolean
              example: true
            balance:
              type: number
              example: 0
    ConnectWalletRequest:
      type: object
      required:
        - blockchain
        - walletAddress
        - signature
        - message
      properties:
        blockchain:
          type: string
          enum: [solana, ethereum]
          example: "solana"
        walletAddress:
          type: string
          example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
        signature:
          type: string
          description: Wallet signature for verification
          example: "0x1234..."
        message:
          type: string
          description: Message that was signed
          example: "Connect wallet to Pangea"
    ConnectWalletResponse:
      type: object
      required:
        - success
        - message
        - wallet
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Wallet connected successfully"
        wallet:
          type: object
          properties:
            address:
              type: string
              example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
            blockchain:
              type: string
              enum: [solana, ethereum]
              example: "solana"
    VerifyWalletRequest:
      type: object
      required:
        - walletAddress
        - signature
        - message
        - blockchain
      properties:
        walletAddress:
          type: string
          example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
        signature:
          type: string
          description: Wallet signature
          example: "0x1234..."
        message:
          type: string
          description: Message that was signed
          example: "Verify wallet ownership"
        blockchain:
          type: string
          enum: [solana, ethereum]
          example: "solana"
    VerifyWalletResponse:
      type: object
      required:
        - success
        - verified
        - wallet
      properties:
        success:
          type: boolean
          example: true
        verified:
          type: boolean
          example: true
        wallet:
          type: object
          properties:
            address:
              type: string
              example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
            blockchain:
              type: string
              enum: [solana, ethereum]
              example: "solana"
    SyncBalancesResponse:
      type: object
      required:
        - success
        - message
        - balances
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Balances synced successfully"
        balances:
          type: array
          items:
            type: object
            description: Balance information for each wallet
    TransactionsResponse:
      type: object
      required:
        - success
        - transactions
      properties:
        success:
          type: boolean
          example: true
        transactions:
          type: array
          items:
            type: object
            properties:
              transactionId:
                type: string
                example: "tx-123"
              hash:
                type: string
                example: "0xabc123..."
              from:
                type: string
                example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
              to:
                type: string
                example: "9wzDvK8vK8vK8vK8vK8vK8vK8vK8vK8vK8vK8vK8vK"
              amount:
                type: number
                example: 10.5
              tokenSymbol:
                type: string
                example: "SOL"
              timestamp:
                type: string
                format: date-time
              status:
                type: string
                enum: [pending, confirmed, failed]
                example: "confirmed"
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
          example: "Invalid credentials"
        error:
          type: string
          description: Error type
          example: "Unauthorized"
        statusCode:
          type: integer
          description: HTTP status code
          example: 401
        details:
          type: object
          description: Additional error details
          additionalProperties: true